初始版本

+++ src/test_mem/test_broker.h
@@ -38,6 +38,7 @@ namespace co {
         int batch_no_index_ = 0;
         std::mutex mutex_;
         std::shared_ptr<std::thread> rep_thread_ = nullptr;
-        std::unordered_map<std::string, std::pair<int, void*>> all_req_;
+        std::unordered_map<std::string, std::pair<int64_t, void*>> all_req_;
+
     };
 }
